refactor: 무한스크롤 로직 분리, 정렬 값 쿼리스트링으로 동작하도록 수정
