+++ app/search/SearchHeader.tsx
@@ -1,6 +1,6 @@
-import { Separator } from "@/components/ui/separator";
-import SubHeader from "@/components/ui/sub-header";
-import { SortType } from "@/types/sort";
+import { Separator } from '@/components/ui/separator';
+import SubHeader from '@/components/ui/sub-header';
+import { SortType } from '@/types/sort';
 
 type SearchHeaderType = {
   current?: SortType;
@@ -11,7 +11,7 @@ export default function SearchHeader({ current, onChange }: SearchHeaderType) {
   return (
     <SubHeader>
       <div className="flex items-center gap-2 max-w-screen-xl w-full overflow-x-scroll scrollbar-hide">
-        <SortButton isCurrent={!current?.orderCol} onClick={() => onChange({ orderCol: "", orderSort: "" })}>
+        <SortButton isCurrent={!current?.orderCol} onClick={() => onChange({ orderCol: '', orderSort: '' })}>
           기본
         </SortButton>
         <Separator orientation="vertical" className="border-r h-4" />
@@ -37,7 +37,7 @@ const SortButton = ({
   return (
     <button
       {...rest}
-      className={`${isCurrent ? "bg-blue-500 text-white" : "bg-transparent text-black"} px-4 py-1 hover:opacity-60 rounded-full border text-nowrap text-caption1 font-normal`}
+      className={`${isCurrent ? 'bg-blue-500 text-white' : 'bg-transparent text-black'} px-4 py-1 hover:opacity-60 rounded-full border text-nowrap text-caption1 font-normal`}
     >
       {children}
     </button>
@@ -46,7 +46,7 @@ const SortButton = ({
 
 type DataListType = { label: string; value: SortType }[];
 const dataList: DataListType = [
-  { label: "많은 가맹점 순", value: { orderCol: "frcsCnt", orderSort: "desc" } },
-  { label: "많은 매출 순", value: { orderCol: "avrgSlsAmt", orderSort: "desc" } },
-  { label: "적은 창업금액 순", value: { orderCol: "smtnAmt", orderSort: "asc" } },
+  { label: '많은 가맹점 순', value: { orderCol: 'frcsCnt', orderSort: 'desc' } },
+  { label: '많은 매출 순', value: { orderCol: 'avrgSlsAmt', orderSort: 'desc' } },
+  { label: '적은 창업금액 순', value: { orderCol: 'smtnAmt', orderSort: 'asc' } },
 ];
